// scripts/types/DappRadar/Dapp.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface Dapp {
    success: boolean;
    error:   null;
    data:    Data;
}

export interface Data {
    info:      Info;
    links:     null;
    contracts: Contract[];
}

export interface Contract {
    title:   string;
    address: string;
}

export interface Info {
    id:                       number;
    balance:                  number;
    author:                   string;
    disabled:                 string;
    title:                    string;
    description:              string;
    description_render:       string;
    url:                      string;
    deepLink:                 string;
    category:                 string;
    featured:                 number;
    hourlyDauGraphUrl:        string;
    volumeLastDay:            number;
    volumeLastWeek:           number;
    volume24hChange:          string;
    txLastDay:                number;
    txLastWeek:               number;
    dauLastDay:               number;
    dauLastWeek:              string;
    weeklyUsers:              string;
    dau24hChange:             string;
    logo:                     string;
    icon:                     string;
    video:                    string;
    updatedAt:                string;
    createdAt:                string;
    nsfw:                     number;
    protocols:                string[];
    images:                   Image[];
    volumeLastDayInUSD:       number;
    volumeLastWeekInUSD:      number;
    volumeLastDayLabel:       string;
    volumeLastWeekLabel:      string;
    volumeLastDayInUSDLabel:  string;
    volumeLastWeekInUSDLabel: string;
    balanceLabel:             string;
    txLastDayLabel:           string;
    txLastWeekLabel:          string;
    dauLastDayLabel:          string;
    slug:                     string;
    isNew:                    boolean;
    changes:                  Changes;
    rankings:                 Rankings;
}

export interface Changes {
    dau: Dau;
    vol: Dau;
}

export interface Dau {
    status: string;
    label:  string;
}

export interface Image {
    title: string;
    url:   string;
}

export interface Rankings {
    overall:  number;
    category: number;
}